MLBEDSW-10414: Port Vela supported-operator checks (2)
Add the following supported operator checks:
ConstraintWeightsPrecision
    Weights must be in 8-bit precision
ConstraintWeightSum
    Validate weight sum to avoid overflow
ConstraintBias
    Bias tensors must be 1D
    Bias tensors must be constant
    Bias tensors must have 32- or 64-bit precision
    Constrain bias range for 64-bit precision

+++ ethosu/regor/test/util.hpp
@@ -61,7 +61,7 @@ std::shared_ptr<Tensor> CreateTensor(const std::string &name, const Shape &stora
 // Create a Const Tensor
 template<typename T>
 std::shared_ptr<Tensor> CreateTensor(const std::string &name, const Shape &storageShape, DataType dtype, std::vector<T> &&values);
-std::shared_ptr<Tensor> CreateTensor(const std::string &name, const Shape &storageShape, DataType dtype, unsigned value);
+std::shared_ptr<Tensor> CreateTensor(const std::string &name, const Shape &storageShape, DataType dtype, int64_t value);
 // Create a Operation with unary input
 std::shared_ptr<Operation> CreateOperation(OpType opType, TensorUsage ifmUsage, std::shared_ptr<Tensor> &ifm,
     TensorUsage ofmUsage, std::shared_ptr<Tensor> &ofm);
